+++ server/src/etl/__init__.py
@@ -1,0 +1,1 @@
+from .etl_client import EtlClient