+++ server/src/routers/action.py
@@ -1,18 +1,22 @@
+import datetime
 from typing import List, Optional
 
 from fastapi import Depends, HTTPException
 from sqlalchemy.orm import Session
 
+from src import settings
 from src.app import app
 
 from src.crud import ActionCrud
 from src.database import get_db
+from src.etl import EtlClient
 from src.schemas import (
     ActionSchema,
     ActionSearchSchema,
     ActionCreateSchema,
     ActionUpdateSchema,
 )
+from src.schemas.action import ActionExportDataSchema
 from src.settings import Settings, get_settings
 
 crud = ActionCrud()
@@ -77,12 +81,30 @@ async def create(data: ActionCreateSchema, db: Session = Depends(get_db)):
     summary=f"Export {prefix}",
 )
 async def export(
+    id_: int,
+    data: ActionExportDataSchema,
     db: Session = Depends(get_db),
 ):
     """
     Export an action
+    :param id_: Action ID
+    :param data: Data
     :param db: Database session
     """
+    action = await crud.get_by_id(db=db, id_=id_)
+    if not action:
+        raise HTTPException(status_code=404, detail=f"{prefix.title()} not found")
+
+    etl_client = EtlClient.from_env()
+    etl_client.export_data(db=db, data=data)
+
+    await crud.update(
+        db=db,
+        id_=id_,
+        data=ActionUpdateSchema(
+            data_exported_at=datetime.datetime.now(),
+        ),
+    )
 
 
 @app.get(